CLI refactoring
