+++ rai/node/node.hpp
@@ -630,8 +630,6 @@ public:
 	void join ();
 	std::vector<std::thread> threads;
 };
-void add_node_options (boost::program_options::options_description &);
-bool handle_node_options (boost::program_options::variables_map &);
 class inactive_node
 {
 public:
